Adds an optional fixed foldlevel for docblock
Adds an option to specify a fixed foldlevel for docblocks.
Having the docblock of the same level as their methods is not
practical. If we want to open all methods with zm the docblocks
will be opened too.
I see three reasons to open a docblock :
    - Write it
    - Edit it
    - Check on what the method does (first reading of a file)
In the daily uses, I prefer to have them fold.