imp: better handling of const and attributes
As the comment says I used a really high fold level for my phpdoc
This way I can quickly unfold everything but the phpdoc
With PHP 7.4 we can now type the attributes, this case was not handle
and was messing up my folds.

In addition I tried to improve a bit the regex that was used by
splitting it in two for const/attribute.
I'm still not a huge fan of the \k since it depends on the user
configuration, I should fix that eventually